Made root dispatcher (dispatcher.ArchiveService) look for web/templates/root.html template and expose service list via data_servicelist.
render_serviceURL and render_ifprotected methods added for that.  There
has to be a better way to do this kind of thing.

New _fc_attr field computer on resource descriptors.  It's horrible and
shouldn't be used.  I don't.

New _fc_property field computer to retrieve DataDescriptor properties
or, failing that, ResourceDescriptor properties.

Changed schema of service list, adapted servicelist.py.  This is now a
horribly overengineered multi-table approach that strongly suggested the 
following changes to the importer framework.

New class table.UniqueForcedTable for importing data sets with multiple
tables, some of which contain rows referring to multiple rows in others
(like the service table).  Corresponding attribute forceUnique on 
RecordDefs.

New option --system to gavoimp.  This makes all shared tables defined
in the RD unshared.  The purpose of this is that you can declare
SharedRecords in system tables and thus use these RDs in everyday life
and create them with the --system option.  This is also a slight
protection against overwriting them accidentally.

New property virtual of data descriptors to inhibit creation of DB
tables for them.  This is done to define tables only used for quering.

New attribute fieldsFrom=id on XML Record elements.  This looks up the
object with id and gets the (initial) fields by calling that object's
get_items method, copies all fields and sets source=dest.  This is what
you want when getting data from a rowsetgrammar.  No idea if it's going
to be useful elsewhere.

Fields can now copy other fields through the "original" XML attribute.

New function makeRowsetFromDicts to create a rowset form a list of
dictionaries (it's a major pain to have to switch between the two all
the time -- I guess it's time we looked into the dict db interface).


git-svn-id: http://svn.ari.uni-heidelberg.de/svn/gavo/python/trunk/docs@153 ca62dbe9-ed84-400c-aab7-804ce7e0b8cc