Built infrastructure for hand-edited registry records.
Sanitized and documented constraints.  They now can come in both
Grammars and Records and can be configured to be fatal or just ignoring.

New macro enterValue.

New @-expansion sourceDate

New config item stateDir for files having to do with the state of the
whole system.  Policy: Nothing should break if it's purged.

if stateDir/MAINT exists, all queries are rejected with a maintenance
notice.


git-svn-id: http://svn.ari.uni-heidelberg.de/svn/gavo/python/trunk/docs@178 ca62dbe9-ed84-400c-aab7-804ce7e0b8cc