New renderer parameterStyle "pql", in use in the dal.xml renderer.
PQL now knows how to make PQL keys from normal input keys, and
standardcores now knows hot to use this.

Moved vizierexprs to svcs -- it cannot be in protocols since svcs needs
to know about it.

Moved pql to svcs, for the same reason as with vizierexprs.

Renamed addRmkFunc to addProcDefObject




git-svn-id: http://svn.ari.uni-heidelberg.de/svn/gavo/python/trunk/docs@1642 ca62dbe9-ed84-400c-aab7-804ce7e0b8cc