Added role 'dachsref' to simplify dachs-ref documentation link.
The new role, 'dachsref' is meant to alias the base-url used when
linking to Dachs-Ref elements documentation. Default base-url is
GAVO DaCHS's one.
The default value is changed when building the docs with sphinx;
currently changing to ReadTheDocs url.
The use of the new role is tested in 'tutorial.rstx' at line 3060
for 'Custom Product Descriptor Generators'.